+++ app-mf-rsbuild/rsbuild.config.ts
@@ -23,10 +23,10 @@ export default defineConfig({
         'rslib_provider': 'mf_provider@http://localhost:4000/mf_provider.js',
         'rslib_manifest_provider': 'mf_provider@http://localhost:4000/mf-manifest.json',
         'rsbuild_provider': 'rsbuild_mf_components@http://localhost:4070/mf-manifest.json',
-        'rsbuild_manifest_provider': 'rsbuild_mf_components@http://localhost:4070/remoteEntry.js',
-        // 'vite_mf_components': 'vite_mf_components@http://localhost:4060/remoteEntry.js',
-        // 'vite_mf_manifest_provider': 'vite_mf_components@http://localhost:4060/mf-manifest.json',
-        'mf_wp': 'remote_mf_webpack@http://localhost:4080/remoteEntry.js', //  will be websoket error, worked
+        'rsbuild_manifest': 'rsbuild_mf_components@http://localhost:4070/remoteEntry.js',
+        // 'vite_mf_remoteEntry': 'vitemfc@http://localhost:4050/remoteEntry.js',
+        'vite_mf_manifest': 'vitemfc@http://localhost:4050/mf-manifest.json',
+        'webpack_mf_remoteEntry': 'remote_mf_webpack@http://localhost:4080/remoteEntry.js', //  will be websoket error, worked
         // 'mf_wp_origin': 'eipReactCommonAssets@http://localhost:9999/remoteEntry.js', //  cant't work
       },
       // shareStrategy: 'loaded-first',