+++ app-mf-vite/src/App.tsx
@@ -9,7 +9,9 @@ import MFRsbuildManifestRsbuild from 'rsbuild_mf_manifest'
 // @ts-ignore
 import RemoteRsbuildProvider from 'mf_rsbuild_remoteEntry'
 // @ts-ignore
-import MFVite from 'vite_mf_manifest'
+import MFViteManifest from 'vite_mf_manifest'
+// @ts-ignore
+import MFViteRemoteEntry from 'vitemfc'
 // @ts-ignore
 import MFWebpack from 'webpack_remoteEntry'
 // @ts-ignore
@@ -20,7 +22,6 @@ function App() {
 
   return (
     <>
-      <Welcome />
       <div>from rslib mf</div>
       <Provider />
       <div>from rslib Manifest mf</div>
@@ -29,8 +30,10 @@ function App() {
       <RemoteRsbuildProvider />
       <div>from rsbuild manifest mf</div>
       <MFRsbuildManifestRsbuild />
-      <div>from vite mf ( can't be work) </div>
-      <MFVite />
+      <div>from vite manifest mf </div>
+      <MFViteManifest />
+      <div>from vite remote entry mf (cant't work)</div>
+      <MFViteRemoteEntry />
       <hr />
       <div>from webpack mf</div>
       <MFWebpack componentName="KYEmpty" />