+++ app-mf-webpack/src/App.tsx
@@ -6,9 +6,9 @@ import React from 'react';
 // @ts-ignore
 // const MFWebpackLazy = React.lazy(() => import('provider'))
 // @ts-ignore
-// const MFViteLazy = React.lazy(() => import('provider'))
+const MFViteLazy = React.lazy(() => import('provider'))
 
-const MFWebpackLazy = React.lazy(() => import('provider/DymicPublicComponent')) // domain worked 
+// const MFWebpackLazy = React.lazy(() => import('provider/DymicPublicComponent')) // domain worked 
 // import MFWebpack from 'provider/DymicPublicComponent' // workd
 const App: React.FC = () => {
 
@@ -19,8 +19,8 @@ const App: React.FC = () => {
             {/* <MFWebpack /> */}
             {/* worked, must use like this */}
             <React.Suspense fallback={<div>Loading...</div>}>
-                <MFWebpackLazy componentName="KYEmpty" />
-                {/* <MFViteLazy /> */}
+                {/* <MFWebpackLazy componentName="KYEmpty" /> */}
+                <MFViteLazy />
             </React.Suspense>
             {/* <MFWebpack componentName="KYEmpty" /> */}
         </div>