+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 2025年5月22日
 
-- [ ] lib-mf-webpack 经过调整,导出组件需要在前面加一个.或./, 才能正常导出模块, 现app-mf-webpack 已正常使用,待测试其他app,待完善readme.md
-- [ ] app-mf-webpack 新增了许多案例,待补充到 readme.md
+- [x] lib-mf-webpack 经过调整,导出组件需要在前面加一个.或./, 才能正常导出模块, 现app-mf-webpack 已正常使用,待测试其他app,待完善readme.md
+- [x] app-mf-webpack 新增了许多案例,待补充到 readme.md
+- [x] 现在已经完成各个打包工具的相互访问,待完善 readme 测试说明